Remove mongo references
