+++ docker-compose.sample.yml
@@ -5,10 +5,8 @@ services:
     build: .
     ports:
       - "5000:5000"
-    links:
-      - "mongodb"
     environment:
-    # replace the following envvars with yours
+    # replace the following env-vars with yours
       PRIVATE_GITHUB_SIGNING_KEY: |
         -----BEGIN RSA PRIVATE KEY-----
         MIIEpQIBAAKCAQEA7FZ0DtQarhPZIxGj5ZH3PlMBf9DoaZOnuXXPb2yD6QiTFH1h
@@ -38,7 +36,7 @@ services:
         Xr6gdKXTH+oo8RTvBqhepvW5GvTaaPZS9wZKzey7oUp6ItJPUEEZmSA=
         -----END RSA PRIVATE KEY-----
       PERIL_INTEGATION_ID: 1234
-      DB_URL: mongodb:27017
+      DATABASE_URL: localhost:5432
       WEB_URL: http://danger.systems
   ngrok:
     image: wernight/ngrok
@@ -47,5 +45,3 @@ services:
     command: "ngrok http peril:5000"
     ports:
       - "4040:4040"
-  mongodb:
-    image: mongo:3.0
