+++ package.json
@@ -24,7 +24,6 @@
     "@types/express": "^4.0.34",
     "@types/jest": "^19.2.2",
     "@types/jsonwebtoken": "^7.1.33",
-    "@types/mongodb": "^2.1.41",
     "@types/node-fetch": "^1.6.6",
     "@types/pg": "^6.1.34",
     "@types/pg-promise": "^0.0.28",
