+++ tsconfig.json
@@ -9,7 +9,8 @@
         ],
         "sourceMap": true,
         "rootDir": "source",
-        "allowJs": true
+        "allowJs": true,
+        "strictNullChecks": true
     },
     "exclude": [
         "source/danger/_tests/fixtures",
